♻️ Refacor: 유저 복구 컴포넌트 분리 (#207)
