+++ src/components/auth/user-recover-modal/index.ts
@@ -1,0 +1,11 @@
+import UserRecoverModal from '@/components/auth/user-recover-modal/UserRecoverModal'
+import UserRecoverCompleteModal from '@/components/auth/user-recover-modal/UserRecoverCompleteModal'
+import UserRecoverFormModal from '@/components/auth/user-recover-modal/UserRecoverFormModal'
+import UserRecoverNoticeModal from '@/components/auth/user-recover-modal/UserRecoverNoticeModal'
+
+export {
+  UserRecoverModal,
+  UserRecoverCompleteModal,
+  UserRecoverFormModal,
+  UserRecoverNoticeModal,
+}
